Inovacao/Chatbot Propostas/2025/model/requirements.txt
flask
flask-cors
openai
python-dotenv
requests
pydantic